change endpoint name and add basic functionality
